fedup.commandline: treat --network VERSION as a float, not int
In our distant past and possible future, distro version numbers are
floating point numbers. We should handle that case.

(See commit 053c9859d62106b3bbbd45eeb65159d8753f4089, which was dropped
when merging with master)

+++ fedup/commandline.py
@@ -161,9 +161,9 @@ def VERSION(arg):
         return 'rawhide'
 
     distro, version, id = platform.linux_distribution()
-    version = int(version)
+    version = float(version)
 
-    if int(arg) >= version:
+    if float(arg) >= version:
         return arg
     else:
         msg = _("version must be greater than %i") % version
